Refactored files to Single Hangman creating an instance of Hangman to play the game, Added replay option when games end.
